# src/gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "UrbanPulse API Gateway",
    "version": "1.0.0",
    "description": "API Gateway para UrbanPulse - Plataforma de movilidad urbana"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Registrar nuevo usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente"
          },
          "400": {
            "description": "Error en validación"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Iniciar sesión",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso"
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        }
      }
    },
    "/api/movilidad/realtime": {
      "get": {
        "tags": ["Tráfico"],
        "summary": "Obtener datos de tráfico en tiempo real",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Datos de tráfico"
          }
        }
      }
    },
    "/api/meteorologia/realtime": {
      "get": {
        "tags": ["Calidad del Aire"],
        "summary": "Obtener calidad del aire en tiempo real",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Datos de calidad del aire"
          }
        }
      }
    }
  }
}